Add mirror app
This app configures the OVS bridge on all the cluster hosts to mirror
traffic from the given VLANs to the interface on the given VM using GRE
tunnels between OVS bridges. RSTP is enabled on each OVS bridge to
prevent loops. If no VLANs are given, then all traffic on the bridge is
mirrored to the GRE tunnel.

Scenario spec should look like the following:

```
spec:
  apps:
  - name: mirror
    hosts:
    - hostname: wireshark
      metadata:
        interface: IF0
        vlans:
        - EXP_1
        - EXP_5
```

Assuming the above scenario config, `EXP_1 --> 101` and `EXP_5 --> 105`,
`IF0 --> mega_tap100`, and the IP of the host the `wireshark` VM is
scheduled on is `10.0.0.105`, the following OVS commands will be run for
the `post-start` stage.  Note that the key is randomly generated.

On all the cluster hosts except for the one the `wireshark` VM is
scheduled on, the following commands will be run:

```
ovs-vsctl set bridge phenix rstp-enable=true
ovs-vsctl add-port phenix wireshark \
  -- set interface wireshark type=gre options:remote_ip=10.0.0.105 options:key=1234567890
ovs-vsctl \
  -- --id=@i0 get port mega_tapABC \
  ...
  -- --id=@i9 get port mega_tapXYZ \
  -- --id=@p get port wireshark \
  -- --id=@m create mirror name=m0 select-dst-port=@i0,...,@i9 select-vlan=101,105 output-port=@p \
  -- set bridge phenix mirrors=@m
```

On the cluster host the `wireshark` VM is scheduled on, the following
commands will be run:

```
ovs-vsctl set bridge phenix rstp-enable=true
ovs-vsctl add-port phenix wireshark \
  -- set interface wireshark type=gre options:remote_ip=flow options:key=1234567890
ovs-ofctl add-flow phenix "in_port=wireshark actions=output:mega_tap100"
ovs-vsctl \
  -- --id=@i0 get port mega_tapABC \
  ...
  -- --id=@i9 get port mega_tapXYZ \
  -- --id=@p get port mega_tap100 \
  -- --id=@m create mirror name=m0 select-dst-port=@i0,...,@i9 select-vlan=101,105 output-port=@p \
  -- set bridge phenix mirrors=@m
```

For the cleanup stage, the following OVS commands are run on all cluster
hosts:

```
ovs-vsctl -- --id=@m get mirror wireshark -- remove bridge phenix mirrors @m
ovs-ofctl del-flows phenix in_port=wireshark
ovs-vsctl del-port phenix wireshark
```

The key to all of this is the `ovs-ofctl add-flow` command run on the
cluster host the wireshark VM is sheduled on. Adding this flow rule
helps to avoid loops created by the GRE tunnels used for mirroring.

**NOTE**: the `mirror` app is currently not namespace aware, and will
fail miserably if two different experiments use the same name for a VM
that mirrored traffic is to be sent to. This will be addressed in a
future version of the app.