Update mirror app to support direct GRE tunnels to target VMs
This update adds a `DirectGRE` option to the mirror app metadata to
support using direct GRE tunnels to target VMs that need to see mirrored
traffic. This is to overcome the network loop issues experienced when
using the original version of this app.

Using the `DirectGRE` option will cause OVS mirrors configured on each
cluster host to mirror traffic over GRE directly to the experiment VM
configured to received the mirrored traffic, as opposed to mirroring
traffic over GRE to another OVS switch. This means the mirrored packets
will still have the GRE header when they reach the target VM.

Most traffic monitoring software supports processing mirrored packets
that include a GRE header. For monitoring software that does not support
mirroring over GRE (or ERSPAN), assuming the monitoring software is
running in Linux, an OVS bridge can be used to terminate the GRE tunnel
in the VM. OVS will then take care of stripping the GRE headers from the
mirrored packets, and the monitoring software can then capture packets
from the `gre_sys` interface created by OVS.

An example scenario spec is as follows:

```
spec:
  apps:
  - name: mirror
    metadata:
      directGRE:
        enabled: true
        mirrorNet: 172.30.0.0/16 // this is the default if not set
        mirrorVLAN: mirror       // this is the default if not set
        mirrorBridge: phenix     // this is the default if not set
    hosts:
    - hostname: wireshark
      metadata:
        interface: IF0
        vlans:
        - EXP_1
        - EXP_5
```

Given the above mirror app config, and assuming the experiment name is
`foobar`, the following will happen:

A tap named `foobar-mirror-app-tap` will be created on each cluster host
on VLAN `foobar//mirror`, using bridge `phenix`, with an IP address
starting at the beginning of the mirror network and incrementing up. So,
if there are 4 hosts that make up the cluster, the following taps will
exist: `172.30.0.[1-4]/16`.

The `wireshark` host in the topology will have it's `IF0` interface
configured with IP address `172.30.255.254/16` on bridge `phenix` and
VLAN `mirror`. If `IF0` doesn't exist on the `wireshark` host it will be
added. If it does exist and has an IP, bridge, and VLAN assigned in the
topology the IP, bridge, and VLAN will be overwritten. If additional
hosts are configured for the app, their IP addresses will increment down
from the end of the mirror network.

A GRE tunnel interface named `foobar-wireshark` will be configured on
each cluster host, with `172.30.255.254` configured as the remote IP
address, and a mirror named `foobar-wireshark` will be configured to use
the GRE tunnel interface as the output port.

By default, a simple GRE tunnel is used. If ERSPAN is required, it can
be enabled via additional app metadata options. As an example:

```
spec:
  apps:
  - name: mirror
    metadata:
      directGRE:
        enabled: true
        mirrorNet: 172.30.0.0/16 // this is the default if not set
        mirrorVLAN: mirror       // this is the default if not set
        mirrorBridge: phenix     // this is the default if not set
        erspan:
          enabled: true
          version: 1
          index: 1
    hosts:
    - hostname: wireshark
      metadata:
        interface: IF0
        vlans:
        - EXP_1
        - EXP_5
```

Both ERSPAN versions 1 and 2 are supported. Version 1 expects the
`index` option to be set, while version 2 expects the `direction` and
`hwid` options to be set.