Add support for Windows to protonuke app
If a protonuke app host is a Windows host, inject a Windows Startup
script to schedule and run protonuke.exe with the args specified in the
host metadata.

The script generated assumes protonuke is present at
C:\minimega\protonuke.exe, which is where it gets injected when the
`phenix image inject-miniexe` subcommand is used to add protonuke to a
Windows image.