[mirror] Automatically avoid mirror net overlap with running experiments
Check for mirror app config in status of other running experiments and
adjust the mirror net for the current experiment if another running
experiment is already using the same subnet.

Also, to prevent from having to manually clean up the cluster in cases
where an experiment didn't stop cleanly, and thus the mirror app's
cleanup stage didn't get executed, we added some cleanup code to the
pre-start stage to ensure taps and mirrors are cleaned up properly. This
will help prevent the cluster from being in a state where taps still
exist with IP addresses to be used by the experiment being started.