feat: new helics app
A new `helics` app was added that looks for a `helics/federate`
annotation on topology nodes and uses those annotation details to
generate HELICS broker init scripts.

Other apps are expected to add the `helics/federate` annotation to
topology nodes that are acting as HELICS federates, and the annotation
details should include the broker endpoint the federate expects to
connect to, as well as the number of federates that will be running on
the node.

```
annotations:
  helics/federate:
    - broker: 172.16.0.25:24000
      fed-count: 1
```